Revert "MODBULKOPS-334 Fix are u sure"
This reverts commit 0fe27cb25d7d1af97cf47cbbde8782e2a4ce28b7.
