+++ src/main/java/org/folio/bulkops/exception/RuleValidationTenantsException.java
@@ -1,14 +1,7 @@
 package org.folio.bulkops.exception;
 
-import lombok.Getter;
-
-@Getter
 public class RuleValidationTenantsException extends Exception {
-
-  private final String identifier;
-
-  public RuleValidationTenantsException(String message, String identifier) {
+  public RuleValidationTenantsException(String message) {
     super(message);
-    this.identifier = identifier;
   }
 }
